moved the extraction of version and description to the tests folder
